New theme released. Please see readme.md
The theme has been updated to be more performant and easier to customise.

To address performance (which was gauged with startuptime.vim), the decision was made to move away from Lush.nvim. Whilst great to get started with, it was taking c. 22ms to load on my machine. Moving away from Lush to a leaner Lua template (thanks to Nightfox.nvim), this has fallen to about 6ms.

Customisation has also been made easier to apply by allowing for theme options to be passed to the theme as a Lua table. This has allowed for styles and highlight groups to be overriden with a lot more clarity and control.

Appreciate this commit breaks things for people and sincerest apologies. I hadn't expected to move away from Lush with this theme and you can still access the previous version on the Lush branch. I'll be accepting pull requests if you do not wish to move to the new version.