fix: allow additional colors to be overridden
The theme has two color palettes; core and additional. The latter is based off of core colors and is used for items such as indent markings, selections, menus etc.

With the previous implementation, if any of the core colors were overriden, the additional colors would not be updated to reflected the new color. This was because they were hard coded with hex values.

In this fix, we use the core color variables instead of hex's.

Signed-off-by: olimorris <olimorris@users.noreply.github.com>