fix: stop double loading of theme for big speed gains!
This theme was based off of the wonderful `EdenEast/nightfox.nvim` theme.

However during refactoring, I noted that the theme would be loaded twice at startup. The impact of this was all of the highlight groups in the theme being loaded, cleared and then loaded again. The impact of this was about 6ms. Loading times after this fix are next to nothing according to my benchmarks.