feat: #23 make more like vscode's onedarkpro
sinkyl made a number of solid suggestions in issue #23 (https://github.com/olimorris/onedarkpro.nvim/issues/23) regarding the theme's lack of compliance to the original onedarkpro.

This commit address _some_ of their concerns regarding highlight groups and theme defaults. As of now, all options are turned off by default. In retrospect, I consider them to be too opinionated to be the default for theme. Also, a number of changes to hlgroups in the `treesitter.lua` and `theme.lua` file have been made as per #23. Finally, the background color is now set at #282c34 for the dark theme.