feat: turn on option styles by default
For a long time, bolds, italics and underlines were turned off by default. To my knowledge, very few had turned them on. Now filetype highlights are turned on by default, this means a lot of users aren't seeing the full potential of filetype highlights. My assumption is that users of this colorscheme are more likely to turn something off than turn it on. By making this change, I've removed all of the opinionated styling from the supported plugins. My view here is that we have a duty (as outlined in opening paragraph of the README.md) to try and match VS Code's OneDarkPro. This means ensuring that syntactical elements match as closely as possible. Applying a bold to the title of `NeoTree` because it looks good was a step too far in terms of expressing one's opinion, in retrospect. As always, open to people's views on this