refactor!: auto-caching and faster load times
Closes #133 #131 #130 and hopefully #112

This PR represents a big leap forward in the theme, thanks to the excellent work of [nightfox.nvim](https://github.com/EdenEast/nightfox.nvim) and [catppuccin](https://github.com/catppuccin/nvim).

As the Neovim community continues to pursue faster startup times, the principle of caching becomes paramount. A previous issue with how the theme implemented caching was the need it placed on the user to remember to refresh the cache. It was hugely onerous.

The summary of the updates in this PR:

- The four themes which make up this colorscheme are now compiled to Lua bytecode and cached to disk
- The bytecode is then executed when the `vim.cmd([[colorscheme onedark]])` is executed
---
- The themes are auto-compiled when either the user's configuration or this plugin, changes
- This is done through some genius work by [Null Chilly](https://github.com/nullchilly) who had the original idea to hash a user's config to disk
- The colorscheme uses "fingerprinting" (a genius [EdenEast](https://github.com/EdenEast) idea) to determine if its codebase has changed and write that to disk
- Comparisons between the computed and disk values then determine if a new cache is required
---
- The principle of generating (or compiling) the themes is pretty much the same as when we were applying highlights in the previous version
- The difference now is that it gets converted to Lua bytecode instead of staying as raw Lua.
---
I've been testing this for about a week now. It's super fast. Each colorscheme loads in around `0.5ms`, down from `0.9ms` with the previous cache implementation and down from about `4ms` without caching.

Some minor improvements include:

- Can specify colors by the background color of the theme; not just by the name
- Each theme exists in its own `.vim` file making it easy to switch themes with external plugins
- More robust test cases and enhanced coverage